Move flattenFixtureTree to shared2 package
