+++ packages/react-cosmos-shared2/src/fixtureTree/createFixtureTree/collapseDirs.test.ts
