+++ packages/react-cosmos-shared2/src/fixtureTree/createFixtureTree/collapseDirs.ts
@@ -1,5 +1,5 @@
 import { forEach } from 'lodash';
-import { FixtureNode, FixtureNodeDirs } from './shared/types';
+import { FixtureNode, FixtureNodeDirs } from '../shared/types';
 
 export function collapseDirs(
   treeNode: FixtureNode,
