+++ packages/react-cosmos-shared2/src/fixtureTree/createFixtureTree/collapseSoloIndexes.test.ts
