+++ packages/react-cosmos-shared2/src/fixtureTree/createFixtureTree/collapseSoloNamedItems.test.ts
