+++ packages/react-cosmos-shared2/src/fixtureTree/createFixtureTree/collapseSoloNamedItems.ts
@@ -1,5 +1,5 @@
 import { forEach } from 'lodash';
-import { FixtureNode, FixtureNodeDirs } from './shared/types';
+import { FixtureNode, FixtureNodeDirs } from '../shared/types';
 
 export function collapseSoloNamedItems(treeNode: FixtureNode): FixtureNode {
   const items = { ...treeNode.items };
