+++ packages/react-cosmos-shared2/src/fixtureTree/createFixtureTree/createRawFixtureTree.ts
@@ -1,6 +1,6 @@
 import { get, set } from 'lodash';
-import { FixtureId, FixtureNamesByPath } from '../renderer';
-import { FixtureNode } from './shared/types';
+import { FixtureId, FixtureNamesByPath } from '../../renderer';
+import { FixtureNode } from '../shared/types';
 
 export function createRawFixtureTree(
   fixtures: FixtureNamesByPath
