+++ packages/react-cosmos-shared2/src/fixtureTree/createFixtureTree/hideFixtureSuffix.test.ts
