+++ packages/react-cosmos-shared2/src/fixtureTree/createFixtureTree/hideSingleChildDirs.test.ts
