+++ packages/react-cosmos-shared2/src/fixtureTree/flattenFixtureTree.test.ts
